requirements.txt
arrow==1.2.1
asgiref==3.4.1
Django==3.2.9
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
factory-boy==3.2.1
Faker==9.8.2
flake8==4.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.2
